--- src/main/java/janggi/infra/transaction/TransactionRunnable.java
@@ -1,8 +1,0 @@
-package janggi.infra.transaction;
-
-import java.sql.Connection;
-
-@FunctionalInterface
-public interface TransactionRunnable {
-    void run(Connection connection) throws Exception;
-}
