--- src/main/java/janggi/infra/transaction/TransactionSupplier.java
@@ -1,8 +1,0 @@
-package janggi.infra.transaction;
-
-import java.sql.Connection;
-
-@FunctionalInterface
-public interface TransactionSupplier<T> {
-    T get(Connection connection) throws Exception;
-}
