+++ src/main/java/janggi/jdbc/dao/RowMapper.java
@@ -1,0 +1,9 @@
+package janggi.jdbc.dao;
+
+import java.sql.ResultSet;
+import java.sql.SQLException;
+
+@FunctionalInterface
+public interface RowMapper<T> {
+    T mapRow(ResultSet rs) throws SQLException;
+}
