+++ src/main/java/janggi/jdbc/dao/game/GameDao.java
@@ -1,0 +1,12 @@
+package janggi.jdbc.dao.game;
+
+import java.sql.Connection;
+import java.util.Optional;
+
+public interface GameDao {
+    Long save(Connection con, String currentTurn);
+
+    Optional<GameEntity> findLatestGame(Connection con);
+
+    void deleteByGameId(Connection con, Long gameId);
+}
