+++ src/main/java/janggi/jdbc/dao/game/GameEntity.java
@@ -1,0 +1,6 @@
+package janggi.jdbc.dao.game;
+
+public record GameEntity(
+        Long id,
+        String currentTurn
+) {}
