+++ src/main/java/janggi/jdbc/dao/piece/PieceEntity.java
@@ -1,0 +1,10 @@
+package janggi.jdbc.dao.piece;
+
+public record PieceEntity(
+        Long id,
+        Long gameId,
+        String pieceType,
+        int positionRow,
+        int positionColumn,
+        String team
+) { }
