+++ src/main/java/janggi/jdbc/transaction/action/TransactionRunnable.java
@@ -1,0 +1,8 @@
+package janggi.jdbc.transaction.action;
+
+import java.sql.Connection;
+
+@FunctionalInterface
+public interface TransactionRunnable {
+    void run(Connection connection);
+}
