+++ src/main/java/janggi/jdbc/transaction/action/TransactionSupplier.java
@@ -1,0 +1,8 @@
+package janggi.jdbc.transaction.action;
+
+import java.sql.Connection;
+
+@FunctionalInterface
+public interface TransactionSupplier<T> {
+    T get(Connection connection);
+}
