--- src/main/java/janggi/repository/dto/LatestInProgressGameResponse.java
@@ -1,8 +1,0 @@
-package janggi.repository.dto;
-
-import janggi.model.Janggi;
-
-public record LatestInProgressGameResponse(
-        Long gameId,
-        Janggi janggi
-) {}
