--- src/main/java/janggi/repository/entity/GameEntity.java
@@ -1,6 +1,0 @@
-package janggi.repository.entity;
-
-public record GameEntity(
-        Long id,
-        String currentTurn
-) {}
