--- src/main/java/janggi/repository/entity/ParameterMapper.java
@@ -1,10 +1,0 @@
-package janggi.repository.entity;
-
-import java.sql.PreparedStatement;
-import java.sql.SQLException;
-
-@FunctionalInterface
-public interface ParameterMapper {
-
-    void map(PreparedStatement preparedStatement) throws SQLException;
-}
