--- src/main/java/janggi/repository/entity/PieceEntity.java
@@ -1,10 +1,0 @@
-package janggi.repository.entity;
-
-public record PieceEntity(
-        Long id,
-        Long gameId,
-        String pieceType,
-        int positionRow,
-        int positionColumn,
-        String team
-) { }
