--- src/main/java/janggi/repository/entity/RowMapper.java
@@ -1,9 +1,0 @@
-package janggi.repository.entity;
-
-import java.sql.ResultSet;
-import java.sql.SQLException;
-
-@FunctionalInterface
-public interface RowMapper<T> {
-    T mapRow(ResultSet rs) throws SQLException;
-}
