+++ src/main/java/janggi/service/dto/LatestInProgressGameResponse.java
@@ -1,0 +1,8 @@
+package janggi.service.dto;
+
+import janggi.model.Janggi;
+
+public record LatestInProgressGameResponse(
+        Long gameId,
+        Janggi janggi
+) {}
