--- src/main/java/janggi/view/Parser.java
@@ -1,12 +1,0 @@
-package janggi.view;
-
-public final class Parser {
-
-    public static int parse(String target) {
-        try {
-            return Integer.parseInt(target);
-        } catch (NumberFormatException e) {
-            throw new IllegalArgumentException("정수가 아닙니다.");
-        }
-    }
-}
