+++ src/main/java/janggi/view/util/Parser.java
@@ -1,0 +1,12 @@
+package janggi.view.util;
+
+public final class Parser {
+
+    public static int parse(String target) {
+        try {
+            return Integer.parseInt(target);
+        } catch (NumberFormatException e) {
+            throw new IllegalArgumentException("정수가 아닙니다.");
+        }
+    }
+}
