+++ src/test/java/janggi/dao/DatabaseTest.java
@@ -1,0 +1,31 @@
+package janggi.dao;
+
+import janggi.config.AppConfig;
+import janggi.jdbc.dao.game.GameDao;
+import janggi.jdbc.dao.game.JdbcGameDao;
+import janggi.jdbc.dao.piece.JdbcPieceDao;
+import java.sql.Connection;
+import java.sql.DriverManager;
+import java.sql.SQLException;
+import org.junit.jupiter.api.AfterEach;
+import org.junit.jupiter.api.BeforeEach;
+
+public abstract class DatabaseTest {
+    protected GameDao gameDao = new JdbcGameDao();
+    protected JdbcPieceDao pieceEntityDao = new JdbcPieceDao();
+    protected Connection con;
+
+    @BeforeEach
+    void beforeEach() throws SQLException {
+        con = DriverManager.getConnection(
+                new AppConfig("application-test.properties").getDbUrl()
+        );
+        con.setAutoCommit(false);
+    }
+
+    @AfterEach
+    void afterEach() throws SQLException {
+        con.rollback();
+        con.close();
+    }
+}
